[PR](laser component): Add laser

+++ src/client/network/NetworkClient.cpp
@@ -265,7 +265,7 @@ std::uint32_t NetworkClient::latencyMs() const noexcept {
     return connection_.latencyMs();
 }
 
-bool NetworkClient::sendInput(std::uint8_t inputMask) {
+bool NetworkClient::sendInput(std::uint16_t inputMask) {
     if (!isConnected() || !serverEndpoint_.has_value() || !socket_->isOpen()) {
         return false;
     }
